+++ meshapp/src/home/logged-out/home.tsx
@@ -5,20 +5,12 @@ import loggedOutNav from "../../components/home-logged-out/loggedOutNavBar";
 import welcomeMessage from "../../components/home-logged-out/loggedOutWelcome";
 import advertSection from "../../components/home-logged-out/advertSection";
 import reviewsSection from "../../components/home-logged-out/reviewSection";
-import LoginInput from "../../components/login-form";
 import { deepmerge } from "@mui/utils";
-import { Login } from "@mui/icons-material";
-import ForgotPassword from "../../components/password-forms/initiate-reset-form";
+import LoginWindow from "../../components/login-form";
 
 export default function LoggedOutPage() {
   const pageTheme = PageTheme();
   
-  const [showForgotPassword, setShowForgotPassword] = useState(false);
-
-  const updateShowForgotPasswordState = () => {
-    setShowForgotPassword(prevState => !prevState)
-  };
-    
   return (
     <ThemeProvider
       theme={(theme: Theme) => {
@@ -67,7 +59,7 @@ export default function LoggedOutPage() {
               filter: "drop-shadow(4px 4px 3px rgba(0,0,0,.6))",
             }}
           >
-            {showForgotPassword ? <ForgotPassword updateShowForgotPasswordState={updateShowForgotPasswordState} /> : <LoginInput updateShowForgotPasswordState={updateShowForgotPasswordState} />}
+            <LoginWindow/>
           </Grid>
         </Grid>
